requirements.txt
imageio==2.21.2
librosa==0.9.2
lmdb==1.3.0
numpy==1.23.5
opencv_python==4.7.0.72
psutil==5.9.4
pyrender==0.1.45
pytorch_lightning==1.9.0
rich==13.5.2
scipy==1.9.1
soundfile==0.11.0
tqdm==4.64.1
transformers==4.26.0
trimesh==3.9.24
diffusers==0.11 # very important 
matplotlib
